fix: Move FAKE_MIGRATION out of test-data
Attempting to depend on anything inside test-utils at runtime is a
footgun, and may result in the server not booting.

This patch stops the server from depending on something that might not
exist at runtime, such as test-utils/test-data trying to get the buffer
of the eamusement csv and failing catastrophically.