fix: make all tests pass
I'm sorry for how disgustingly large all of these commits are. I promise
that this is the last time something this disgusting will happen.

This is a mono-patch for various things that were broken by the major
refactor. It's mostly things like error messages now being different
(and tests failing as a result).

Some notable fixes include me accidentally deleting the entire DELETE
api-tokens endpoint (lol), and timeAchieved: 0 in batch-manual not
coercing to null anymore. It appears the behaviour was intentional for
some reason. I'm not sure why though.