chore: make certain @types not devDeps
These are actually necessary in non-dev scenarios, as we use
ts-node/register as our loader. This expects types to exist
for our dependencies, for some reason.

I'm too tired to properly resolve this, so making @types mandatory
is my hack fix.

It's not like they add much to the build size anyway.