feat: Experimental go re-implementation of sieglinde
THIS CODE DOES NOT WORK! IT NEVER CONVERGES, AND I HAVEN'T INVESTIGATED
WHY JUST YET.

However, this is an experiemental go re-impl of sieglinde for
performance.
Sadly, it doesn't seem to uh, do much, for performance. I'm leaving it
here as a commit for clarities sake, but it isn't going to be used for
anything nor kept in the repo.