docs: enforce pnpm run <scriptname>
Can you believe it? pnpm <scriptname> goes through this bizarre
anti-logic wrt. running scripts.

If the script is a dedicated pnpm command, it will run that command.
Else if the script exists in package.json, it will run it.

At this point, it then tries to run it as a normal shell command. By
checking if <scriptname> is in $PATH and then executing that.

What the hell? This was helpfully revealed to my by a tester who found
`pnpm sort` was taking too long. That's because it was running
`sort(1)`.

Absolutely insane behaviour. Why???