fix: remove more instances of instanceof
it's completely broken and no longer honoured. I *still* to this day do
not know why this is a bug on our systems. Maybe I will investigate it
further at some point, but for now, instanceof is banned in the codebase
as it doesn't seem to fucking work.

normally i refrain from swearing in a codebase i know future employers
will check, but this has to be one of *the* most frustrating bugs we've
had in years.