Remove Arcaea gauge
It turns out that Arcaea's implementation of gauge is intensely stupid,
because the maximum gauge value is based on which partner is selected
(e.g. Toa Kozukata's max gauge is 60).

It's rare that anyone would care about the remaining life at the end of
the chart anyways, unless it's something like a 0-gauge HARD CLEAR.