refactor: port github-bot to ESM
This is one of those annoying things. Updating OctoKit updated some transitive sindresorhus dependency which now forces onto ESM.

This proves to be a pretty annoying problem because you can't use CommonJS from inside ESM and vice versa. Somehow, as an industry, we've managed to invent "what color is your import" for NodeJS, which is a pretty brutal problem.

Anyway. I've made this specific package into an ESM module, and what it compiles to is a module now, too, instead of CommonJS. Luckily, because the github bot is so small and isolated, it means we have very little trouble doing this (and hopefully no brutal side effects).