remove newline from MyPageScraper CSV parser check (#1207)
It seems there are CSVs in the wild that use \r\n instead of just \n, which can cause this to fail. Since this check is already best-effort, just drop the newline since it doesn't really matter.

Untested. Unclear if the \r\n can cause additional downstream  serverside parsing issues that will also need to be addressed.