seed: IIDX INFINITAS Song Pack Vol. 27 (#1304)
* seed: IIDX INFINITAS Song Pack Vol. 27

+30 EPOLIS songs from in AC

And the rest of the songs from special selection pack will be commit later.
I set this PR as draft for now. (Just to make sure I'm not do something wrong).

I will told you when it ready to merge.

* add 9(?) new INFINITAS exclusive songs.

- WILD FIRE	がるる
- VIVID DEBUT!	#EmoCosine
- Astrogazer	DJ TOTTO
- STAY GOLD	Yuta Imai
- 大見解	Des-ROW・組
- Inside Insight	BEMANI Sound Team "HuΣeR" feat.Fernweh
- タンポポ	豚乙女
- REFLEXES MANIPULATION	山本真央樹
- onslaught	PHQUASE

Only missing is Caldera because it's using different ingameID and it's on the seed. I don't know is it possible to insert that in the old one.

(midnight and my London skill is going bad lol. )

* added CALDERA

Thank you zk!